Add cast em valor em round.

+++ src/ingest-step1-ini.sql
@@ -434,8 +434,8 @@ CREATE or replace FUNCTION ingest.feature_asis_assign_volume(
             WHEN 'line'  THEN round( ST_Length(geom,true)/1000.0)::int
             ELSE  null::int
           END size,
-          round( ST_Area(ST_OrientedEnvelope(geom),true)/1000000, 1)::int AS bbox_km2,
-          round(size_mdn,3) AS size_mdn
+          round( (ST_Area(ST_OrientedEnvelope(geom),true)/1000000)::numeric, 1)::int AS bbox_km2,
+          round(size_mdn::numeric,3) AS size_mdn
         FROM (
             SELECT count(*) n, st_collect(geom) as geom, CASE gtype --(gtype||iif(p_usemedian,'','_no'::text))
                 WHEN 'poly'  THEN percentile_disc (0.5) WITHIN GROUP(ORDER BY ST_Area(geom,true)) /1000000.0
